requirements.txt
typer==0.3.0
typer-cli==0.0.9

PyYAML==5.3.1

rich==2.2.6

flit==2.3.0

mypy==0.782
isort==4.3.21
flake8==3.8.3
flake8-blind-except==0.1.1
flake8-broken-line==0.2.0
flake8-bugbear==20.1.4
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.2.3
flake8-debugger==3.2.1
flake8-eradicate==0.4.0
flake8-isort==3.0.0
flake8-mock==0.3
flake8-mutable==1.2.0
flake8-pytest==1.3
flake8-quotes==3.2.0
flake8-string-format==0.3.0
flake8-variables-names==0.0.3
